+++ oo/src/language/running/environment.rs
@@ -1,37 +1,59 @@
 use std::collections::HashMap;
 
-use crate::language::running::value::{Binding, Value};
+use crate::language::{analyzing::stree::STree, running::value::{Binding, Value}};
 
+
+pub struct Scope {
+    pub bindings: HashMap<String, Binding>,
+    pub deferred: Vec<STree>
+}
 
 pub struct Environment {
-    scopes: Vec<HashMap<String, Binding>>
+    scopes: Vec<Scope>,
 }
 
 impl Environment {
     pub fn new() -> Environment {
-        Environment { scopes: vec![HashMap::new()] }
+        Environment { 
+            scopes: vec![
+                Scope {
+                    bindings: HashMap::new(),
+                    deferred: Vec::new()
+                }
+            ]
+        }
     }
 
     pub fn push_scope(&mut self) {
-        self.scopes.push(HashMap::new());
+        self.scopes.push(Scope { bindings: HashMap::new(), deferred: Vec::new() });
     }
 
-    pub fn pop_scope(&mut self) {
+    pub fn peek_scope(&mut self) -> Option<&Scope> {
         if self.scopes.len() > 1 {
-            self.scopes.pop();
+            self.scopes.last()
+        } else {
+            None
+        }
+    }
+
+    pub fn pop_scope(&mut self) -> Option<Scope> {
+        if self.scopes.len() > 1 {
+            self.scopes.pop()
+        } else {
+            None
         }
     }
 
     pub fn declare(&mut self, name: String, value: Value, mutable: bool) {
         if let Some(scope) = self.scopes.last_mut() {
-            scope.insert(name, Binding::new(value, mutable));
+            scope.bindings.insert(name, Binding::new(value, mutable));
         }
     }
 
 
     pub fn set(&mut self, name: &str, value: Value) -> Result<(), String> {
         for scope in self.scopes.iter_mut().rev() {
-            if let Some(binding) = scope.get_mut(name) {
+            if let Some(binding) = scope.bindings.get_mut(name) {
                 return binding.set(value);
             }
         }
@@ -41,7 +63,7 @@ impl Environment {
 
     pub fn get(&self, name: &str) -> Result<Value, String> {
         for scope in self.scopes.iter().rev() {
-            if let Some(binding) = scope.get(name) {
+            if let Some(binding) = scope.bindings.get(name) {
                 return Ok(binding.value().clone());
             }
         }
@@ -50,11 +72,17 @@ impl Environment {
 
     pub fn is_mutable(&self, name: &str) -> Result<bool, String> {
         for scope in self.scopes.iter().rev() {
-            if let Some(binding) = scope.get(name) {
+            if let Some(binding) = scope.bindings.get(name) {
                 return Ok(binding.mutable.clone());
             }
         }
         Err(format!("Variable '{}' not found", name))
     }
     
+    pub fn defer(&mut self, stmt: STree) {
+        if let Some(scope) = self.scopes.last_mut() {
+            scope.deferred.push(stmt);
+        }
+    }
+
 }