+++ oo/CHANGELOG.md
@@ -32,7 +32,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
-- Immutable Variables 
+- Immutable variables 
+- Empty returns for null functions
+- Defer statements
 
 ### Changed